Update job treatment

+++ src/php/admin/job/treatmentEditJob.php
@@ -1,5 +1,4 @@
 <?php
-var_dump($_POST);
 // SI L'IMAGE N'A PAS ÉTÉ CHANGÉE //
 if ($_SERVER['REQUEST_METHOD'] == 'POST' && $_FILES['update_picture_job']['name'] == '') {
 
@@ -52,7 +51,8 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST' && $_FILES['update_picture_job']['name'
                 $job->setJobResponsabilities($selectedResponsabilities);
 
                 // Update en db
-                var_dump($job);die;
+                $jobRepository = new JobRepository();
+                $jobRepository->updateJob($job);
 
                 $_SESSION['success-message'] = "L'offre d'emploie à bien été modifier !";
                 header('location: index.php?admin&action=AdminEmplois');
@@ -74,25 +74,53 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST' && $_FILES['update_picture_job']['name'
         exit();
     }
 
-} elseif ($_SERVER['REQUEST_METHOD'] == 'POST' && $_FILES['']['name'] !== '') {
+} elseif ($_SERVER['REQUEST_METHOD'] == 'POST' && $_FILES['update_picture_job']['name'] !== '') {
+    
+    if (isset($_POST['update_id_job']) && isset($_POST['update_title_job']) && isset($_POST['update_desc_job']) && isset($_FILES['update_picture_job']) && isset($_POST['update_chief_job']) && isset($_POST['update_date_job_created']) && isset($_POST['update_date_job_started']) && isset($_POST['selectedLocations']) && isset($_POST['selectedQualifications']) && isset($_POST['selectedResponsabilities'])) {
+    
+        if ($_POST['update_id_job'] !== '' && $_POST['update_title_job'] !== '' && $_POST['update_desc_job'] !== '' && $_FILES['update_picture_job'] !== '' && $_POST['update_chief_job'] !== '' && $_POST['update_date_job_created'] !== '' && $_POST['update_date_job_started'] !== '' && $_POST['selectedLocations'] !== '' && $_POST['selectedQualifications'] !== '' && $_POST['selectedResponsabilities'] !== '') {
 
+            // Stockage et filtrage
+            $jobId = $_POST['update_id_job'];
+            $titleJob = htmlspecialchars(strip_tags(trim($_POST['update_title_job'])));
+            $descJob = htmlspecialchars(strip_tags(trim($_POST['update_desc_job'])));
+            $chiefJob = htmlspecialchars(strip_tags(trim($_POST['update_chief_job'])));
+            $selectedLocations = $_POST['selectedLocations'];
+            $selectedQualifications = $_POST['selectedQualifications'];
+            $selectedResponsabilities = $_POST['selectedResponsabilities'];
 
-    if (isset($_POST['']) && isset($_POST[''])) {
-
-        if ($_POST[''] !== '' && $_POST[''] !== '') {
+            // Dates
+            $dateCreated = htmlspecialchars(strip_tags(trim($_POST['update_date_job_created'])));
+            $dateStarted = htmlspecialchars(strip_tags(trim($_POST['update_date_job_started'])));
+            // Convertir les dates en objets DateTime
+            $dateCreatedObj = DateTime::createFromFormat('Y-m-d\TH:i', $dateCreated);
+            $dateStartedObj = DateTime::createFromFormat('Y-m-d\TH:i', $dateStarted);
+            // Vérifier si les dates sont valides
+            if (!$dateCreatedObj || !$dateStartedObj) {
+                $_SESSION['error-message'] = "Les dates fournies ne sont pas valides !";
+                header('location: index.php?admin&action=AdminEditEmploi');
+                exit();
+            }
+            // Vérifier si la date de début est postérieure ou égale à la date de création
+            if ($dateStartedObj < $dateCreatedObj) {
+                $_SESSION['error-message'] = "La date de début doit être postérieure ou égale à la date de création !";
+                header('location: index.php?admin&action=AdminEditEmploi');
+                exit();
+            }
+            // Formater les dates dans le format DATETIME
+            $formattedDateCreated = $dateCreatedObj ? $dateCreatedObj->format('Y-m-d H:i:s') : null;
+            $formattedDateStarted = $dateStartedObj ? $dateStartedObj->format('Y-m-d H:i:s') : null;
 
             try {
 
                 // On va chercher le lien de l'image
-                $postRepository = new PostRepository();
-                $image = 1;
-                $imagePath = 1;
+                $jobRepository = new JobRepository();
+                $job = $jobRepository->findJobById($jobId);
+                $imagePath = $job->getJobPicture();
+
 
                 // On prépare l'url du fichier à supprimer
-                $basePath = '/'; // Remplacez cette valeur par le chemin absolu de votre projet
                 $path = $imagePath;
-
-                // $path = $_SERVER['DOCUMENT_ROOT'] . '/localhost/' . $imagePath;
 
                 // On tente de supprimer le fichier et quittons le script si il n'est pas supprimer et stockons une erreur en session
                 if (!unlink($path)) {
@@ -101,13 +129,13 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST' && $_FILES['update_picture_job']['name'
                     header('location: index.php?admin&action=AdminEmplois');
                     exit;
 
-                } else {
+                }
 
-                    $nameFile = $_FILES['']['name'];
-                    $typeFile = $_FILES['']['type'];
-                    $tmpFile = $_FILES['']['tmp_name'];
-                    $errorFile = $_FILES['']['error'];
-                    $sizeFile = $_FILES['']['size'];
+                    $nameFile = $_FILES['update_picture_job']['name'];
+                    $typeFile = $_FILES['update_picture_job']['type'];
+                    $tmpFile = $_FILES['update_picture_job']['tmp_name'];
+                    $errorFile = $_FILES['update_picture_job']['error'];
+                    $sizeFile = $_FILES['update_picture_job']['size'];
 
                     $extensions = ['png', 'jpg', 'jpeg', 'gif', 'jiff'];
                     $type = ['image/png', 'image/jpg', 'image/jpeg', 'image/gif', 'image/jiff'];
@@ -121,11 +149,27 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST' && $_FILES['update_picture_job']['name'
                         if (count($extension) <= 2 && in_array(strtolower(end($extension)), $extensions)) {
                             if ($sizeFile <= $max_size && $errorFile == 0) {
                                 if (move_uploaded_file($tmpFile, $image = 'upload/' . uniqid() . '.' . end($extension))) {
+ 
+                                    // Instancier un objet Job et définir ses propriétés
+                                    $job = new Job();
+                                    $job->setJobId($jobId);
+                                    $job->setJobTitle($titleJob);
+                                    $job->setJobDescription($descJob);
+                                    $job->setJobChiefName($chiefJob);
+                                    $job->setJobDateCreated($formattedDateCreated);
+                                    $job->setJobDateStarted($formattedDateStarted);
+                                    $job->setJobPlaces($selectedLocations);
+                                    $job->setJobQualifications($selectedQualifications);
+                                    $job->setJobResponsabilities($selectedResponsabilities);
 
-                                    // Instancier un objet Post et définir ses propriétés
+                                    // Update en db
+                                    $jobRepository = new JobRepository();
+                                    $jobRepository->updateJob($job);
 
-                                    // On supprime et réinserre les valeurs
-
+                                    //Update image et description de la table job :)
+                                    $jobRepository->updateJobImage($jobId, $image, $nameFile);
+                                    
+                               
                                     $_SESSION['success-message'] = "Votre offre d'emploie a bien été modifié !";
                                     header('location: index.php?admin&action=AdminEmplois');
                                     exit();
@@ -152,7 +196,7 @@ if ($_SERVER['REQUEST_METHOD'] == 'POST' && $_FILES['update_picture_job']['name'
                         exit();
                     }
 
-                }
+                
 
             } catch (Exception $e) {
                 die('Erreur : ' . $e->getMessage());
